Fix: 두개로 나눌 필요 없어보이는 뷰 합체 !
